ngx_http_log_module.c: Reduce diff to upstream
Allowing a full capabilty for ngx_http_log_copy_short is suboptimal as
it requires a call to memcpy that can't be inlined. The memcpy call in
ngx_http_log_copy_short is also dangerous and could overflow the source
without suboject bounds (if assertions are disabled), and enforcing
bounds clutters the code. Instead, use the same code as upstream, just
with ngx_vaddr_t rather than uintptr_t as the data type.