[cheri] Make the ngx_http_ephemeral_t union explicit.
The ngx_http_request_t structure contains an implicit union between a
chunk of fields that are no longer used after parsing the request
line, and the ngx_ephemeral_t structure.
This patch makes the union explicit so that sub-object bounds work
correctly. Note that we retain the old layout as well, so that
compiler warnings and static analysis tools can be tested against it.