[cheri] Do not set bounds for regex capture data.
The ngx script subsystem can run a regex on headers and other data.
The capture groups, as well as other variables are associated with the
request as a set of variables. These variables can be referenced in
filters and other modules by name.
The regex capture groups break with aggressive sub-object bounds
because a 1-byte capability is used instead of the full string length.
Note: this should ensure that the length is minimal, I'm not sure this
is currently the case.